requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.3.0
bert-score==0.3.13
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
cmake==4.1.0
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.1
ema-pytorch==0.7.7
evalplus==0.2.1
evaluate==0.4.6
exceptiongroup==1.3.0
fastapi==0.118.0
fasttext-wheel==0.9.2
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
inflect==7.5.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.3.0
lightning-utilities==0.15.2
llvmlite==0.45.1
lm-format-enforcer==0.10.3
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.6
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multipledispatch==1.0.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==2.1.0
outlines==0.0.46
packaging==25.0
pandas==2.2.3
peft==0.15.2
pillow==11.0.0
platformdirs==4.4.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.4.0
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyairports @ https://files.pythonhosted.org/packages/6e/75/b424aebc9f2fc5db319d5df5fff62fa19254c8ef974c254588d48c480df2/pyairports-2.1.1-py3-none-any.whl#sha256=3dfa0cc3e47696692ade92feccdc6b046968f2a75f5e30f65735d6db7251cb26
pyarrow==21.0.0
pybind11==3.0.1
pycountry==24.6.1
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.49.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rouge_score @ https://files.pythonhosted.org/packages/e2/c5/9136736c37022a6ad27fea38f3111eb8f02fe75d067f9a985cc358653102/rouge_score-0.1.2.tar.gz#sha256=c7d4da2683e68c9abf0135ef915d63a46643666f848e558a1b9f7ead17ff0f04
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.39.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.3
tempdir==0.7.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchmetrics==1.8.2
torchvision==0.19.0+cu121
tqdm==4.67.1
transformers==4.57.0
triton==3.0.0
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
vllm==0.5.4
vllm-flash-attn==2.6.1
wandb==0.22.1
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
xformers==0.0.27.post2
xxhash==3.6.0
yarl==1.20.1
